finish method in photo service - save photo

+++ ThereNoLogin/Models/NewTherePlace.swift
@@ -25,8 +25,8 @@ struct NewTherePlace: Identifiable, Codable {
     var placeState: String = ""
     var placeCountry: String = ""
     var placeCountryCode: Int?
-    var latGPS: Float16?
-    var lonGPS: Float16?
+    var latGPS: Double?
+    var lonGPS: Double?
     var dateVisitedStart: Date?
     var dateVisitedEnd: Date?
     var timeVisitedStart: Date?
